refactoring and test
