+++ packages/core/src/interactions/packet/acknowledgement.ts
@@ -5,8 +5,11 @@ import { getUnacknowledgedTickets, deleteTicket, replaceTicketWithAcknowledgemen
 import type HoprCoreEthereum from '@hoprnet/hopr-core-ethereum'
 import { PROTOCOL_ACKNOWLEDGEMENT } from '../../constants'
 import PeerId from 'peer-id'
-import { PublicKey } from '@hoprnet/hopr-utils'
+import { PublicKey, durations, Hash } from '@hoprnet/hopr-utils'
+import type { Packet } from '../../messages/packet'
 const log = debug('hopr-core:acknowledgement')
+
+const ACKNOWLEDGEMENT_TIMEOUT = durations.seconds(2)
 
 export function subscribeToAcknowledgements(
   subscribe: any,
@@ -15,7 +18,7 @@ export function subscribeToAcknowledgements(
   pubKey: PeerId,
   onMessage: (ackMessage: Acknowledgement) => void
 ) {
-  subscribe(PROTOCOL_ACKNOWLEDGEMENT, async function (msg: Uint8Array, remotePeer: PeerId) {
+  subscribe(PROTOCOL_ACKNOWLEDGEMENT, async function (msg: Uint8Array, remotePeer: PeerId) { 
     const ackMsg = Acknowledgement.deserialize(msg, pubKey, remotePeer)
     let unacknowledgedTicket = await getUnacknowledgedTickets(db, ackMsg.ackChallenge)
 
@@ -27,7 +30,7 @@ export function subscribeToAcknowledgements(
 
     const channel = chain.getChannel(new PublicKey(pubKey.pubKey.marshal()), new PublicKey(remotePeer.pubKey.marshal()))
 
-    const ackedTicket = await channel.acknowledge(unacknowledgedTicket, ackMsg.ackChallenge)
+    const ackedTicket = await channel.acknowledge(unacknowledgedTicket, new Hash(ackMsg.ackKeyShare))
 
     if (ackedTicket === null) {
       log(`Got a ticket that is not a win. Dropping ticket.`)
@@ -36,6 +39,17 @@ export function subscribeToAcknowledgements(
       log(`Storing winning ticket`)
       await replaceTicketWithAcknowledgement(db, ackMsg.ackChallenge, ackedTicket)
     }
+
     onMessage(ackMsg)
+  })
+}
+
+export function sendAcknowledgement(packet: Packet, destination: PeerId, sendMessage: any, privKey: PeerId): void {
+  setImmediate(async () => {
+    const ack = packet.createAcknowledgement(privKey)
+
+    sendMessage(destination, PROTOCOL_ACKNOWLEDGEMENT, ack.serialize(), {
+      timeout: ACKNOWLEDGEMENT_TIMEOUT
+    })
   })
 }
